moved expertise logic to hook
